+++ docs/.vitepress/config.ts
@@ -41,8 +41,10 @@ export default defineConfig({
           { text: 'Contract', link: 'contract' },
           { text: 'Channel', link: 'channel' },
           { text: 'Registry', link: 'registry' },
+          { text: 'Broker', link: 'broker' },
           { text: 'Controller', link: 'controller' },
           { text: 'Endpoint', link: 'endpoint' },
+          { text: 'Message', link: 'message' },
         ],
       },
       {
@@ -58,28 +60,6 @@ export default defineConfig({
                 text: 'Overview',
                 link: 'overview',
               },
-              {
-                text: 'Installation',
-                link: 'installation',
-              },
-              {
-                text: '@Avro.InjectSchema',
-                link: 'schema',
-                base: '/decorators/schema/',
-              },
-              {
-                text: '@Avro.Record',
-                link: 'record',
-                base: '/packages/schema-avro/decorators/fields/',
-                collapsed: true,
-                items: [
-                  { text: '@Avro.Int', link: 'int' },
-                  { text: '@Avro.String', link: 'string' },
-                  { text: '@Avro.Double', link: 'double' },
-                  { text: '@Avro.Boolean', link: 'double' },
-                  { text: '🚧 @Avro....', link: 'double' },
-                ],
-              },
             ],
           },
           {
@@ -91,14 +71,6 @@ export default defineConfig({
                 text: 'Overview',
                 link: 'overview',
               },
-              {
-                text: 'Installation',
-                link: 'installation',
-              },
-              {
-                text: '@Confluent.Subject',
-                link: '/decorators/inject-subject',
-              },
             ],
           },
           {
@@ -109,36 +81,6 @@ export default defineConfig({
               {
                 text: 'Overview',
                 link: 'overview',
-              },
-              {
-                text: 'Installation',
-                link: 'installation',
-              },
-              {
-                text: '@Kafka.InjectClient',
-                link: 'schema',
-                base: '/decorators/inject-client',
-              },
-              {
-                text: '@Kafka.InjectTopic',
-                link: 'schema',
-                base: '/decorators/inject-topic',
-              },
-              {
-                text: '@Kafka.Controller',
-                link: 'schema',
-                base: '/decorators/controller/',
-              },
-              {
-                text: '@Kafka.Endpoint',
-                link: 'record',
-                base: '/packages/client-kafka/decorators/params/',
-                collapsed: true,
-                items: [
-                  { text: '@Kafka.Key', link: 'key' },
-                  { text: '@Kafka.Value', link: 'value' },
-                  { text: '@Avro.Meta', link: 'meta' },
-                ],
               },
             ],
           },