+++ examples/t2-nestjs/apps/order/src/order.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, Logger } from '@nestjs/common';
 import { OrderRepository } from './repository/OrderRepository';
 import { OrderItem, OrderStatus } from './repository/OrderItem';
 
@@ -7,14 +7,22 @@ export class OrderService {
   oderRepository: OrderRepository = new OrderRepository();
 
   createOrder(sessionID: string) {
-    const order = new OrderItem(sessionID);
-    return this.oderRepository.save(order);
+    const order = this.oderRepository.save(new OrderItem(sessionID));
+    Logger.log('order created ...', { order });
+    return order;
   }
 
   rejectOrder(orderId: string) {
-    const order = this.oderRepository.findById(orderId);
-    if (!order) throw new Error('Order not found');
-    order.status = OrderStatus.FAILURE;
-    return this.oderRepository.save(order);
+    const orders = this.oderRepository.findAll();
+
+    for (const order of orders) {
+      if (order.sessionID === orderId) {
+        Logger.log('order failed ...', { order });
+        order.status = OrderStatus.FAILURE;
+        return this.oderRepository.save(order);
+      }
+    }
+
+    throw new Error('Order not found');
   }
 }