+++ packages/client/src/index.ts
@@ -2,5 +2,5 @@ export * from './Serializer';
 export * from './AbstractMessage';
 
 export * from './channel';
-export * from './client';
+export * from './broker';
 export * from './controller';