+++ packages/client-kafka/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @ts-messaging/client-kafka
 
+## 1.0.0
+
+### Major Changes
+
+- Initial Research Prototype Release
+
+### Patch Changes
+
+- Updated dependencies
+  - @ts-messaging/client@1.0.0
+  - @ts-messaging/common@1.0.0
+
 ## 0.0.0
 
 ### Patch Changes