+++ packages/client/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @ts-messaging/client
 
+## 1.0.0
+
+### Major Changes
+
+- Initial Research Prototype Release
+
+### Patch Changes
+
+- Updated dependencies
+  - @ts-messaging/common@1.0.0
+
 ## 0.0.0
 
 ### Patch Changes