+++ packages/common/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @ts-messaging/common
 
+## 1.0.0
+
+### Major Changes
+
+- Initial Research Prototype Release
+
 ## 0.0.0
 
 ### Patch Changes